requirements.txt
Flask==2.3.2
pymongo==4.6.0
werkzeug==2.3.7
python-dotenv==1.0.1
dnspython==2.4.2
PyJWT==2.8.0
gunicorn==21.2.0
